Remove CIS Event Service V1 From Subaccount Cleanup Job
