+++ resources/keb/values.yaml
@@ -683,8 +683,6 @@ serviceBindingCleanup:
 subaccountCleanup:
   enabled: "true"
   schedule: "0 1 * * *"
-  # Specifies the Events Service version.
-  eventsServiceVersion: "v2"
 # =================================================
